Reapply "[LLVM] Make the GPU loader utilities an LLVM tool (#132096)"
Summary:
There were a few issues with the first one, leading to some errors and
warnings. Most importantly, this was building on MSVC which isn't
supported.
