--- libc/utils/gpu/loader/nvptx/CMakeLists.txt
@@ -1,9 +1,0 @@
-set(LLVM_LINK_COMPONENTS
-  BinaryFormat
-  Object
-  Option
-  Support
-  )
-
-add_llvm_executable(nvptx-loader nvptx-loader.cpp)
-target_link_libraries(nvptx-loader PRIVATE gpu_loader CUDA::cuda_driver)
